More std fixes

+++ indexer/feature.cpp
@@ -552,7 +552,7 @@ FeatureType::GeomStat FeatureType::GetOuterTrianglesStats()
 {
   ASSERT(!m_parsed.m_triangles, ("Triangles had been parsed already!"));
   CHECK(m_loadInfo, ());
-  size_t const scalesCount = m_loadInfo->GetScalesCount();
+  int const scalesCount = m_loadInfo->GetScalesCount();
   ASSERT_LESS_OR_EQUAL(scalesCount, DataHeader::kMaxScalesCount, ("MWM has too many geometry scales!"));
   FeatureType::GeomStat res;
